+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 2022-10-21
+### Changed
+- JSON fields that have `@(This.Default)` and are set to their default value, are not encoded.
+  This does not apply to fields that are marked with a nonstandard default value, such as `@(This.Default!1)`!
+
 ## [1.6.4] - 2022-10-14
 ### Fixed
 - Support DMD 2.100.